requirements.txt
fastapi
uvicorn[standard]
exa-py
firecrawl-py
x402[fastapi,evm]
python-dotenv
